资源删除，将BigDecimal改为Double

+++ service/service-oss/src/main/java/com/dizhongdi/serviceoss/entity/DzdSource.java
@@ -44,7 +44,7 @@ public class DzdSource implements Serializable {
     private String sourceName;
 
     @ApiModelProperty(value = "文件大小：MB")
-    private BigDecimal fileSize;
+    private Double fileSize;
 
     @ApiModelProperty(value = "md5值")
     private String md5;
